+++ Examples/ISHDeploy/Set-ADFSIntegration.ps1
@@ -7,7 +7,7 @@ param (
     [switch]$IncludeInternalClients=$false
 )        
 
-$ishBootStrapRootPath="C:\GitHub\ISHBootstrap"
+$ishBootStrapRootPath=Resolve-Path "$PSScriptRoot\..\.."
 $cmdletsPaths="$ishBootStrapRootPath\Source\Cmdlets"
 $scriptsPaths="$ishBootStrapRootPath\Source\Scripts"